Update get-activation-key.md
I have spent the last week trying to identify why my storage gateway would not activate inside my Secured VPC which is part of a LandingZone deployment. I have eventually identified that this is related to the requirement to include the private VPCEndpoint in the URL. 
This is not clearly documented anywhere and so I suggest it is added here.